// admcentro-web-main/src/interfaces/IclientPayments.ts
// Generated by https://quicktype.io

import { IPerson } from './Iowners'
import { Iproperty } from './Iproperties'

export interface IClienyPaymentsResponse {
  results: number
  code: number
  status: string
  ok: boolean
  data: IClienyPayment[]
}

export interface IClienyPayment {
  expenseDetails: ExpenseDetail[]
  eventualityDetails: EventualityDetail[]
  id: number
  ContractId: number
  OwnerId?: number
  Owner?: IPerson
  PaymentTypeId: number
  recharge: number
  year: number
  month: string
  total: number
  createdAt: string
  updatedAt: string
  deletedAt: null
  rentingAmount: number
  Contract: Contract
  PaymentType: PaymentType
  printData?: any
  obs?: string
  paidTotal: number
}

export interface Contract {
  id: number
  PropertyId: number
  ClientId: number
  Client: IPerson
  Property: Iproperty
  startDate: string
  endDate: string
  commission: number
  state: string
  amount: number
  description: string
  stamped: number
  fees: number
  warrantyInquiry: number
  createdAt: string
  updatedAt: string
  deletedAt: null
  adjustmentMonth: number
}

export interface PaymentType {
  id: number
  name: string
  createdAt: string
  updatedAt: string
  deletedAt: null
}

export interface EventualityDetail {
  id: number
  ContractId: number
  clientAmount: number
  ownerAmount: number
  clientPaid: boolean
  ownerPaid: boolean
  description: string
  expiredDate: string
  createdAt: string
  updatedAt: string
  deletedAt: null
}

export interface ExpenseDetail {
  id: number
  description: string
  amount: number
  date: string
  ContractId: number
  createdAt: string
  updatedAt: string
  deletedAt: null
}